Needed to change ints to longs to handle billions. Added unit test.

+++ NumberToWordsLib/Handlers/BillionsHandler.cs
@@ -3,12 +3,12 @@
     // Concrete handler for billions
     public class BillionsHandler : NumberHandler
     {
-        public override string Handle(int number)
+        public override string Handle(long number)
         {
             if (number >= 1000000000 && number < 1000000000000)
             {
-                int billions = number / 1000000000;
-                int remainder = number % 1000000000;
+                long billions = number / 1000000000;
+                long remainder = number % 1000000000;
                 return NumberToWordsBelow1000(billions) + " billion" + (remainder > 0 ? " " + Handle(remainder) : "");
             }
             else
