+++ NumberToWordsLib/Handlers/UnitsHandler.cs
@@ -3,7 +3,7 @@
     // Concrete handler for units (less than 1000)
     public class UnitsHandler : NumberHandler
     {
-        public override string Handle(int number)
+        public override string Handle(long number)
         {
             if (number < 1000)
             {
